+++ pkg/core/hooks/conn/time_offset.go
@@ -1,0 +1,43 @@
+//go:build linux
+
+package conn
+
+import (
+	"fmt"
+	"time"
+
+	"golang.org/x/sys/unix"
+)
+
+var (
+	realTimeOffset uint64
+)
+
+// InitRealTimeOffset calculates the offset between the real clock and the monotonic clock used in the BPF.
+func initRealTimeOffset() error {
+	var monotonicTime, realTime unix.Timespec
+	if err := unix.ClockGettime(unix.CLOCK_MONOTONIC, &monotonicTime); err != nil {
+		return fmt.Errorf("failed getting monotonic clock due to: %v", err)
+	}
+	if err := unix.ClockGettime(unix.CLOCK_REALTIME, &realTime); err != nil {
+		return fmt.Errorf("failed getting real clock time due to: %v", err)
+	}
+	realTimeOffset = uint64(time.Second)*(uint64(realTime.Sec)-uint64(monotonicTime.Sec)) + uint64(realTime.Nsec) - uint64(monotonicTime.Nsec)
+	// realTimeCopy := time.Unix(int64(realTimeOffset/1e9), int64(realTimeOffset%1e9))
+	// log.Debug(fmt.Sprintf("%s real time offset is: %v", Emoji, realTimeCopy))
+	return nil
+}
+
+// GetRealTimeOffset is a getter for the real-time-offset.
+func getRealTimeOffset() uint64 {
+	return realTimeOffset
+}
+
+// convertUnixNanoToTime takes a Unix timestamp in nanoseconds as a uint64 and returns the corresponding time.Time
+func convertUnixNanoToTime(unixNano uint64) time.Time {
+	// Unix time is the number of seconds since January 1, 1970 UTC,
+	// so convert nanoseconds to seconds for time.Unix function
+	seconds := int64(unixNano / uint64(time.Second))
+	nanoRemainder := int64(unixNano % uint64(time.Second))
+	return time.Unix(seconds, nanoRemainder)
+}
